refactor: 랜덤 숫자 생성, 입력값 확인 모듈화
